Introduce OwnedIoEventLoop that allows the user to keep control over the EventLoop processing.
Motivation:

In the past we had multiple times the request that some people wanted to own the EventLoop processing. This was not easily doable in 4.1 but now with the seperation of IoHandler and IoEventLoop it is.

Modifications:

Add OwnedIoEventLoop that users can use if they desire to drive the EventLoop themselves.

Result:

Be able to drive EventLoop manually for more advances use-cases